refactor: move EmailValidatorAdapter to main layer
