+++ src/main/factories/signup/SignUpValidationCompositeFactory.ts
@@ -2,7 +2,7 @@ import Validation from '../../../presentation/protocols/Validation'
 import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
 import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
 import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
-import EmailValidatorAdapter from '../../../utils/EmailValidatorAdapter'
+import EmailValidatorAdapter from '../../adapters/EmailValidatorAdapter'
 
 const SignUpValidationCompositeFactory = {
   makeSignUpValidationComposite (): Validation {
